Refactor some functions out into a new file, expr_info.{h,cpp}
