+++ toolchain/check/pattern_match.cpp
@@ -15,6 +15,7 @@
 #include "toolchain/check/subpattern.h"
 #include "toolchain/check/type.h"
 #include "toolchain/diagnostics/format_providers.h"
+#include "toolchain/sem_ir/expr_info.h"
 #include "toolchain/sem_ir/pattern.h"
 
 namespace Carbon::Check {
